Update 04-deploy-app-aks.md
Shouldn't we guide the user to create the kubeconfig file? 
If running the lab from scratch there is no explanation about the "az aks get-credentials" command to generate the kubeconfig file to connect to the cluster. If the user is running  "kubectl create secret" directly he will get an error about the server rejecting the connection (ther is no connection)